+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Styled excel export: added support for numeric column multiplier
 
 ## 1.0.4
 * Styled excel export: added support for valueRender